requirements.txt
# Mango backend runtime dependencies

# Django / API
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
django-filter==25.2
django-cors-headers==4.9.0

# Async tasks / cache
celery==5.6.2
django-redis==6.0.0

# Database
psycopg2-binary==2.9.11
pgvector==0.4.2

# Market / news.txt data
requests==2.32.5
aiohttp==3.12.15
beautifulsoup4==4.13.4
lxml==6.0.0
pandas==3.0.1
yfinance==1.2.0
Pillow==11.1.0
akshare==1.18.27
exchange-calendars==4.13.1

# Config / parsing
python-dotenv==1.2.2
PyYAML==6.0.2

# AI / LLM
dashscope==1.25.15
openai==2.30.0
pydantic==2.12.5
langchain==1.2.14
langchain-classic==1.0.3
langchain-core==1.2.25
langchain-community==0.4.1
langchain-openai==1.1.12
langgraph==1.1.6

# Production web server
gunicorn==23.0.0; platform_system != "Windows"
